Add Prometheus and Grafana to Docker Compose and enable actuator Prometheus endpoint
This update includes adding Prometheus and Grafana services to the docker-compose file and enabling the actuator prometheus endpoint in the application. Additionally, necessary config files for Prometheus and Grafana have been created and the Docker Compose file extension has been corrected from .yaml to .yml for convention. A micrometer dependency for data observation has also been added to the